refactor email-server and job-server
