+++ packages/email-server/.editorconfig
