+++ packages/email-server/.eslintrc.js
