+++ packages/email-server/src/emails/components/Footer.tsx
