+++ packages/email-server/src/index.ts
@@ -1,0 +1,1 @@
+export { SendEmailPayload, EmailTemplateNames, EmailClient } from './client';
