+++ packages/email-server/tsconfig.json
@@ -3,6 +3,7 @@
   "compilerOptions": {
     "jsx": "react-jsx",
     "allowSyntheticDefaultImports": true,
+    "declaration": true,
     "outDir": "dist",
     "sourceMap": true
   },
