+++ packages/job-server/package.json
@@ -22,6 +22,7 @@
     "watch:ts": "tsc -w"
   },
   "dependencies": {
+    "@bull-board/fastify": "^4.9.0",
     "@fastify/autoload": "^5.0.0",
     "@fastify/cookie": "^8.3.0",
     "@fastify/env": "^4.1.0",
@@ -29,13 +30,16 @@
     "@fastify/redis": "^6.1.0",
     "@fastify/sensible": "^5.0.0",
     "@ossinsight/api-server": "workspace:^1.0.0",
+    "@ossinsight/email-server": "workspace:^0.0.1",
     "@types/cron": "^2.0.0",
     "async": "^3.2.4",
+    "axios": "^0.27.2",
     "bufferutil": "^4.0.7",
     "bullmq": "^3.4.1",
     "cron": "^2.1.0",
     "dotenv": "^16.0.3",
     "encoding": "^0.1.13",
+    "fast-glob": "^3.2.12",
     "fastify": "^4.0.0",
     "fastify-cli": "^5.5.1",
     "fastify-cron": "^1.3.1",
