+++ packages/job-server/src/env.ts
@@ -11,8 +11,20 @@ export const JobServerEnvSchema = {
       type: 'string',
       default: 'redis://localhost:6379/0',
     },
+    EMAIL_SERVER_URL: {
+      type: 'string',
+      default: 'http://127.0.0.1:30002/api',
+    },
     PLAYGROUND_DATABASE_URL: {
       type: 'string',
-    }
+    },
+    SEND_REPO_FEEDS_CRON: {
+      type: 'string',
+      default: '50 * * * *',
+    },
+    CALC_REPO_MILESTONES_CRON: {
+      type: 'string',
+      default: '20 * * * *',
+    },
   },
 };
