+++ packages/job-server/src/plugins/redis.ts
@@ -2,7 +2,10 @@ import fp from "fastify-plugin";
 import fastifyRedis from "@fastify/redis";
 
 export default fp(async (app) => {
-    app.register(fastifyRedis, { url: app.config.REDIS_URL });
+    app.register(fastifyRedis, {
+        url: app.config.REDIS_URL,
+        maxRetriesPerRequest: null
+    });
 }, {
     name: 'fastify-redis',
     dependencies: [
